+++ stories/components/table.stories.tsx
@@ -9,12 +9,15 @@ import {
 } from '../../src/components/index.js';
 import { table } from '../data/data.js';
 
+type TableRecord = (typeof table)[number];
+
 interface ControlProps {
   bordered?: boolean;
   compact?: boolean;
   interactive?: boolean;
   striped?: boolean;
   stickyHeader?: boolean;
+  virtualizeRows?: boolean;
 }
 
 export default {
@@ -27,6 +30,7 @@ export default {
     ),
   ],
   args: {
+    virtualizeRows: false,
     bordered: false,
     compact: false,
     interactive: false,
@@ -89,6 +93,7 @@ export function Control({
   interactive,
   striped,
   stickyHeader,
+  virtualizeRows,
 }: ControlProps) {
   return (
     <Table
@@ -98,29 +103,9 @@ export function Control({
       striped={striped}
       stickyHeader={stickyHeader}
       columns={columns}
-      data={table}
-    />
-  );
-}
-
-export function Virtualized({
-  bordered,
-  compact,
-  interactive,
-  striped,
-  stickyHeader,
-}: ControlProps) {
-  return (
-    <Table
-      bordered={bordered}
-      compact={compact}
-      interactive={interactive}
-      striped={striped}
-      stickyHeader={stickyHeader}
-      columns={columns}
-      data={table}
-      virtualizeRows
+      virtualizeRows={virtualizeRows}
       estimatedRowHeight={() => 172}
+      data={table}
     />
   );
 }
@@ -156,6 +141,7 @@ export function CustomHeaderCellRender({
   interactive,
   striped,
   stickyHeader,
+  virtualizeRows,
 }: ControlProps) {
   return (
     <Table
@@ -172,6 +158,8 @@ export function CustomHeaderCellRender({
           width: '100%',
         },
       }}
+      virtualizeRows={virtualizeRows}
+      estimatedRowHeight={() => 172}
       renderHeaderCell={(thProps, header) => {
         const backgroundColor = header.column.columnDef.meta?.color;
         const width = header.column.columnDef.meta?.width;
@@ -182,6 +170,20 @@ export function CustomHeaderCellRender({
           />
         );
       }}
+    />
+  );
+}
+
+export function StyledTable(props: ControlProps) {
+  const StyledTableComponent = styled(Table<TableRecord>)`
+    border: 1px solid red;
+  `;
+  return (
+    <StyledTableComponent
+      {...props}
+      data={table}
+      columns={columns}
+      estimatedRowHeight={() => 172}
     />
   );
 }